+++ view/base/web/js/view/ace-wrapper.js
@@ -2,17 +2,22 @@
  * @package Goomento_PageBuilder
  * @link https://github.com/Goomento/PageBuilder
  */
-define(['!domReady'], function () {
-    require.config({
-        paths: {
-            ace: ['//cdnjs.cloudflare.com/ajax/libs/ace/1.2.5']
-        }
-    })
 
-    setTimeout(() => {
-        require([
-            'ace/ace',
-            'ace/ext-language_tools',
-        ], () => {});
-    }, 100);
+require.config({
+    paths: {
+        ace: ['//cdnjs.cloudflare.com/ajax/libs/ace/1.2.5']
+    }
+});
+
+define([], function () {
+    'use strict';
+
+    require([
+        'ace/ace',
+        'ace/ext-language_tools',
+    ], (ace) => {
+        window.ace = window.ace || ace;
+    });
+
+    return window.ace;
 });