+++ view/base/web/js/view/intro-wrapper.js
@@ -1,0 +1,12 @@
+/**
+ * @package Goomento_PageBuilder
+ * @link https://github.com/Goomento/PageBuilder
+ */
+define([
+    'Goomento_PageBuilder/lib/intro/intro.min'
+], function (introJs) {
+    'use strict';
+
+    window.introJs = window.introJs || introJs;
+    return introJs;
+});